requirements.txt
# مكتبات المشروع
requests==2.31.0          # لإرسال طلبات HTTP
httpx                     # مكتبة HTTP
python-telegram-bot==20.0  # لإدارة بوت تلغرام
transformers
